// src/common/types.ts
export enum CharacterClassNumber {
  /// <summary>
  /// Value for the dark wizard character class.
  /// </summary>
  DarkWizard = 0,

  /// <summary>
  /// Value for the soul master character class.
  /// </summary>
  SoulMaster = 2,

  /// <summary>
  /// Value for the grand master character class.
  /// </summary>
  GrandMaster = 3,

  /// <summary>
  /// Value for the dark knight character class.
  /// </summary>
  DarkKnight = 4,

  /// <summary>
  /// Value for the blade knight character class.
  /// </summary>
  BladeKnight = 6,

  /// <summary>
  /// Value for the blade master character class.
  /// </summary>
  BladeMaster = 7,

  /// <summary>
  /// Value for the fairy elf character class.
  /// </summary>
  FairyElf = 8,

  /// <summary>
  /// Value for the muse elf character class.
  /// </summary>
  MuseElf = 10,

  /// <summary>
  /// Value for the high elf character class.
  /// </summary>
  HighElf = 11,

  /// <summary>
  /// Value for the magic gladiator character class.
  /// </summary>
  MagicGladiator = 12,

  /// <summary>
  /// Value for the duel master character class.
  /// </summary>
  DuelMaster = 13,

  /// <summary>
  /// Value for the dark lord character class.
  /// </summary>
  DarkLord = 16,

  /// <summary>
  /// Value for the lord emperor character class.
  /// </summary>
  LordEmperor = 17,

  /// <summary>
  /// Value for the summoner character class.
  /// </summary>
  Summoner = 20,

  /// <summary>
  /// Value for the bloody summoner character class.
  /// </summary>
  BloodySummoner = 22,

  /// <summary>
  /// Value for the dimension master character class.
  /// </summary>
  DimensionMaster = 23,

  /// <summary>
  /// Value for the rage fighter character class.
  /// </summary>
  RageFighter = 24,

  /// <summary>
  /// Value for the fist master character class.
  /// </summary>
  FistMaster = 25,
}
// ...
export enum PlayerClass {
  DarkWizard = 1,
  SoulMaster = 201,
  GrandMaster = 301,
  SoulWizard = 401,
  DarkKnight = 2,
  BladeKnight = 202,
  BladeMaster = 302,
  DragonKnight = 402,
  FairyElf = 3,
  MuseElf = 203,
  HighElf = 303,
  NobleElf = 403,
  MagicGladiator = 4,
  DuelMaster = 304,
  MagicKnight = 404,
  DarkLord = 5,
  LordEmperor = 305,
  EmpireLord = 405,
  Summoner = 6,
  BloodySummoner = 206,
  DimensionMaster = 306,
  DimensionSummoner = 406,
  RageFighter = 7,
  FistMaster = 307,
  FistBlazer = 407,
  GlowLancer = 8,
  MirageLancer = 308,
  ShiningLancer = 408,
  RuneMage = 9,
  RuneSpellMaster = 209,
  GradRuneMaster = 309,
  MajesticRuneWizard = 409,
  Slayer = 10,
  RoyalSlayer = 210,
  MasterSlayer = 310,
  Slaughterer = 410,
  GunCrusher = 11,
  GunBreaker = 211,
  MasterGunBreaker = 311,
  HeistGunCrasher = 411,
  WhiteWizard = 12,
  LightMaster = 212,
  ShineWizard = 312,
  ShineMaster = 412,
  Mage = 13,
  WoMage = 213,
  ArchMage = 313,
  MysticMage = 413,
  IllusionKnight = 14,
  MirageKnight = 214,
  IllusionMaster = 314,
  MysticKnight = 414,
  Alchemist = 15,
  AlchemicMaster = 215,
  AlchemicForce = 315,
  Creator = 415,
}